Feature gate the BMI2 move making function, refactor.
