+++ src/board/bitboard.rs
@@ -1,5 +1,8 @@
 //! Contains a number of utility functions and constants for bitboard operations.
 
+#[cfg(feature="bmi2")]
+use bitintr::x86::bmi2::{pdep, pext};
+
 /// The 'A' file of the board.
 pub const FILE_A: u64 = 0x80_80_80_80_80_80_80_80;
 /// The 'B' file of the board.
@@ -187,6 +190,297 @@ pub const SHIFT_RAYS: [[u64; 8]; 64] = [
       0x8040201008040200, 0x0000000000000000, 0x0000000000000000, 0x0000000000000000 ],
 ];
 
+/// Rays which are used by the PEXT/PDEP move making function.
+pub const EXTRACT_RAYS: [[u64; 4]; 64] = [
+    [0xff00000000000000, 0x8080808080808080, 0x8040201008040201, 0x8000000000000000],
+    [0xff00000000000000, 0x4040404040404040, 0x4020100804020100, 0x4080000000000000],
+    [0xff00000000000000, 0x2020202020202020, 0x2010080402010000, 0x2040800000000000],
+    [0xff00000000000000, 0x1010101010101010, 0x1008040201000000, 0x1020408000000000],
+    [0xff00000000000000, 0x0808080808080808, 0x0804020100000000, 0x0810204080000000],
+    [0xff00000000000000, 0x0404040404040404, 0x0402010000000000, 0x0408102040800000],
+    [0xff00000000000000, 0x0202020202020202, 0x0201000000000000, 0x0204081020408000],
+    [0xff00000000000000, 0x0101010101010101, 0x0100000000000000, 0x0102040810204080],
+    [0x00ff000000000000, 0x8080808080808080, 0x0080402010080402, 0x4080000000000000],
+    [0x00ff000000000000, 0x4040404040404040, 0x8040201008040201, 0x2040800000000000],
+    [0x00ff000000000000, 0x2020202020202020, 0x4020100804020100, 0x1020408000000000],
+    [0x00ff000000000000, 0x1010101010101010, 0x2010080402010000, 0x0810204080000000],
+    [0x00ff000000000000, 0x0808080808080808, 0x1008040201000000, 0x0408102040800000],
+    [0x00ff000000000000, 0x0404040404040404, 0x0804020100000000, 0x0204081020408000],
+    [0x00ff000000000000, 0x0202020202020202, 0x0402010000000000, 0x0102040810204080],
+    [0x00ff000000000000, 0x0101010101010101, 0x0201000000000000, 0x0001020408102040],
+    [0x0000ff0000000000, 0x8080808080808080, 0x0000804020100804, 0x2040800000000000],
+    [0x0000ff0000000000, 0x4040404040404040, 0x0080402010080402, 0x1020408000000000],
+    [0x0000ff0000000000, 0x2020202020202020, 0x8040201008040201, 0x0810204080000000],
+    [0x0000ff0000000000, 0x1010101010101010, 0x4020100804020100, 0x0408102040800000],
+    [0x0000ff0000000000, 0x0808080808080808, 0x2010080402010000, 0x0204081020408000],
+    [0x0000ff0000000000, 0x0404040404040404, 0x1008040201000000, 0x0102040810204080],
+    [0x0000ff0000000000, 0x0202020202020202, 0x0804020100000000, 0x0001020408102040],
+    [0x0000ff0000000000, 0x0101010101010101, 0x0402010000000000, 0x0000010204081020],
+    [0x000000ff00000000, 0x8080808080808080, 0x0000008040201008, 0x1020408000000000],
+    [0x000000ff00000000, 0x4040404040404040, 0x0000804020100804, 0x0810204080000000],
+    [0x000000ff00000000, 0x2020202020202020, 0x0080402010080402, 0x0408102040800000],
+    [0x000000ff00000000, 0x1010101010101010, 0x8040201008040201, 0x0204081020408000],
+    [0x000000ff00000000, 0x0808080808080808, 0x4020100804020100, 0x0102040810204080],
+    [0x000000ff00000000, 0x0404040404040404, 0x2010080402010000, 0x0001020408102040],
+    [0x000000ff00000000, 0x0202020202020202, 0x1008040201000000, 0x0000010204081020],
+    [0x000000ff00000000, 0x0101010101010101, 0x0804020100000000, 0x0000000102040810],
+    [0x00000000ff000000, 0x8080808080808080, 0x0000000080402010, 0x0810204080000000],
+    [0x00000000ff000000, 0x4040404040404040, 0x0000008040201008, 0x0408102040800000],
+    [0x00000000ff000000, 0x2020202020202020, 0x0000804020100804, 0x0204081020408000],
+    [0x00000000ff000000, 0x1010101010101010, 0x0080402010080402, 0x0102040810204080],
+    [0x00000000ff000000, 0x0808080808080808, 0x8040201008040201, 0x0001020408102040],
+    [0x00000000ff000000, 0x0404040404040404, 0x4020100804020100, 0x0000010204081020],
+    [0x00000000ff000000, 0x0202020202020202, 0x2010080402010000, 0x0000000102040810],
+    [0x00000000ff000000, 0x0101010101010101, 0x1008040201000000, 0x0000000001020408],
+    [0x0000000000ff0000, 0x8080808080808080, 0x0000000000804020, 0x0408102040800000],
+    [0x0000000000ff0000, 0x4040404040404040, 0x0000000080402010, 0x0204081020408000],
+    [0x0000000000ff0000, 0x2020202020202020, 0x0000008040201008, 0x0102040810204080],
+    [0x0000000000ff0000, 0x1010101010101010, 0x0000804020100804, 0x0001020408102040],
+    [0x0000000000ff0000, 0x0808080808080808, 0x0080402010080402, 0x0000010204081020],
+    [0x0000000000ff0000, 0x0404040404040404, 0x8040201008040201, 0x0000000102040810],
+    [0x0000000000ff0000, 0x0202020202020202, 0x4020100804020100, 0x0000000001020408],
+    [0x0000000000ff0000, 0x0101010101010101, 0x2010080402010000, 0x0000000000010204],
+    [0x000000000000ff00, 0x8080808080808080, 0x0000000000008040, 0x0204081020408000],
+    [0x000000000000ff00, 0x4040404040404040, 0x0000000000804020, 0x0102040810204080],
+    [0x000000000000ff00, 0x2020202020202020, 0x0000000080402010, 0x0001020408102040],
+    [0x000000000000ff00, 0x1010101010101010, 0x0000008040201008, 0x0000010204081020],
+    [0x000000000000ff00, 0x0808080808080808, 0x0000804020100804, 0x0000000102040810],
+    [0x000000000000ff00, 0x0404040404040404, 0x0080402010080402, 0x0000000001020408],
+    [0x000000000000ff00, 0x0202020202020202, 0x8040201008040201, 0x0000000000010204],
+    [0x000000000000ff00, 0x0101010101010101, 0x4020100804020100, 0x0000000000000102],
+    [0x00000000000000ff, 0x8080808080808080, 0x0000000000000080, 0x0102040810204080],
+    [0x00000000000000ff, 0x4040404040404040, 0x0000000000008040, 0x0001020408102040],
+    [0x00000000000000ff, 0x2020202020202020, 0x0000000000804020, 0x0000010204081020],
+    [0x00000000000000ff, 0x1010101010101010, 0x0000000080402010, 0x0000000102040810],
+    [0x00000000000000ff, 0x0808080808080808, 0x0000008040201008, 0x0000000001020408],
+    [0x00000000000000ff, 0x0404040404040404, 0x0000804020100804, 0x0000000000010204],
+    [0x00000000000000ff, 0x0202020202020202, 0x0080402010080402, 0x0000000000000102],
+    [0x00000000000000ff, 0x0101010101010101, 0x8040201008040201, 0x0000000000000001]
+];
+
+/// An array containing outflank information for an extracted ray.
+pub const OUTFLANK : [[u8; 64]; 8] =[
+    [
+    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+    0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20,
+    0x10, 0x10, 0x10, 0x10, 0x10, 0x10, 0x10, 0x10, 0x08, 0x08, 0x08, 0x08, 0x04, 0x04, 0x02, 0x01
+    ], [
+    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+    0x10, 0x10, 0x10, 0x10, 0x10, 0x10, 0x10, 0x10, 0x08, 0x08, 0x08, 0x08, 0x04, 0x04, 0x02, 0x01,
+    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+    0x10, 0x10, 0x10, 0x10, 0x10, 0x10, 0x10, 0x10, 0x08, 0x08, 0x08, 0x08, 0x04, 0x04, 0x02, 0x01
+    ], [
+    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x08, 0x08, 0x08, 0x08, 0x04, 0x04, 0x02, 0x01,
+    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x08, 0x08, 0x08, 0x08, 0x04, 0x04, 0x02, 0x01,
+    0x80, 0x80, 0x80, 0x80, 0x80, 0x80, 0x80, 0x80, 0x88, 0x88, 0x88, 0x88, 0x84, 0x84, 0x82, 0x81,
+    0x80, 0x80, 0x80, 0x80, 0x80, 0x80, 0x80, 0x80, 0x88, 0x88, 0x88, 0x88, 0x84, 0x84, 0x82, 0x81
+    ], [
+    0x00, 0x00, 0x00, 0x00, 0x04, 0x04, 0x02, 0x01, 0x00, 0x00, 0x00, 0x00, 0x04, 0x04, 0x02, 0x01,
+    0x40, 0x40, 0x40, 0x40, 0x44, 0x44, 0x42, 0x41, 0x40, 0x40, 0x40, 0x40, 0x44, 0x44, 0x42, 0x41,
+    0x00, 0x00, 0x00, 0x00, 0x04, 0x04, 0x02, 0x01, 0x00, 0x00, 0x00, 0x00, 0x04, 0x04, 0x02, 0x01,
+    0x80, 0x80, 0x80, 0x80, 0x84, 0x84, 0x82, 0x81, 0x80, 0x80, 0x80, 0x80, 0x84, 0x84, 0x82, 0x81
+    ], [
+    0x00, 0x00, 0x02, 0x01, 0x00, 0x00, 0x02, 0x01, 0x20, 0x20, 0x22, 0x21, 0x20, 0x20, 0x22, 0x21,
+    0x00, 0x00, 0x02, 0x01, 0x00, 0x00, 0x02, 0x01, 0x40, 0x40, 0x42, 0x41, 0x40, 0x40, 0x42, 0x41,
+    0x00, 0x00, 0x02, 0x01, 0x00, 0x00, 0x02, 0x01, 0x20, 0x20, 0x22, 0x21, 0x20, 0x20, 0x22, 0x21,
+    0x00, 0x00, 0x02, 0x01, 0x00, 0x00, 0x02, 0x01, 0x80, 0x80, 0x82, 0x81, 0x80, 0x80, 0x82, 0x81
+    ], [
+    0x00, 0x01, 0x00, 0x01, 0x10, 0x11, 0x10, 0x11, 0x00, 0x01, 0x00, 0x01, 0x20, 0x21, 0x20, 0x21,
+    0x00, 0x01, 0x00, 0x01, 0x10, 0x11, 0x10, 0x11, 0x00, 0x01, 0x00, 0x01, 0x40, 0x41, 0x40, 0x41,
+    0x00, 0x01, 0x00, 0x01, 0x10, 0x11, 0x10, 0x11, 0x00, 0x01, 0x00, 0x01, 0x20, 0x21, 0x20, 0x21,
+    0x00, 0x01, 0x00, 0x01, 0x10, 0x11, 0x10, 0x11, 0x00, 0x01, 0x00, 0x01, 0x80, 0x81, 0x80, 0x81
+    ], [
+    0x00, 0x00, 0x08, 0x08, 0x00, 0x00, 0x10, 0x10, 0x00, 0x00, 0x08, 0x08, 0x00, 0x00, 0x20, 0x20,
+    0x00, 0x00, 0x08, 0x08, 0x00, 0x00, 0x10, 0x10, 0x00, 0x00, 0x08, 0x08, 0x00, 0x00, 0x40, 0x40,
+    0x00, 0x00, 0x08, 0x08, 0x00, 0x00, 0x10, 0x10, 0x00, 0x00, 0x08, 0x08, 0x00, 0x00, 0x20, 0x20,
+    0x00, 0x00, 0x08, 0x08, 0x00, 0x00, 0x10, 0x10, 0x00, 0x00, 0x08, 0x08, 0x00, 0x00, 0x80, 0x80
+    ], [
+    0x00, 0x04, 0x00, 0x08, 0x00, 0x04, 0x00, 0x10, 0x00, 0x04, 0x00, 0x08, 0x00, 0x04, 0x00, 0x20,
+    0x00, 0x04, 0x00, 0x08, 0x00, 0x04, 0x00, 0x10, 0x00, 0x04, 0x00, 0x08, 0x00, 0x04, 0x00, 0x40,
+    0x00, 0x04, 0x00, 0x08, 0x00, 0x04, 0x00, 0x10, 0x00, 0x04, 0x00, 0x08, 0x00, 0x04, 0x00, 0x20,
+    0x00, 0x04, 0x00, 0x08, 0x00, 0x04, 0x00, 0x10, 0x00, 0x04, 0x00, 0x08, 0x00, 0x04, 0x00, 0x80
+    ]
+];
+
+/// An array containing flip information given a specific outflank. Works on extracted lines.
+pub const FLIP : [[u8;256]; 8] = [
+    [
+    0x00, 0xfe, 0xfc, 0xfc, 0xf8, 0xf8, 0xf8, 0xf8, 0xf0, 0xf0, 0xf0, 0xf0, 0xf0, 0xf0, 0xf0, 0xf0,
+    0xe0, 0xe0, 0xe0, 0xe0, 0xe0, 0xe0, 0xe0, 0xe0, 0xe0, 0xe0, 0xe0, 0xe0, 0xe0, 0xe0, 0xe0, 0xe0,
+    0xc0, 0xc0, 0xc0, 0xc0, 0xc0, 0xc0, 0xc0, 0xc0, 0xc0, 0xc0, 0xc0, 0xc0, 0xc0, 0xc0, 0xc0, 0xc0,
+    0xc0, 0xc0, 0xc0, 0xc0, 0xc0, 0xc0, 0xc0, 0xc0, 0xc0, 0xc0, 0xc0, 0xc0, 0xc0, 0xc0, 0xc0, 0xc0,
+    0x80, 0x80, 0x80, 0x80, 0x80, 0x80, 0x80, 0x80, 0x80, 0x80, 0x80, 0x80, 0x80, 0x80, 0x80, 0x80,
+    0x80, 0x80, 0x80, 0x80, 0x80, 0x80, 0x80, 0x80, 0x80, 0x80, 0x80, 0x80, 0x80, 0x80, 0x80, 0x80,
+    0x80, 0x80, 0x80, 0x80, 0x80, 0x80, 0x80, 0x80, 0x80, 0x80, 0x80, 0x80, 0x80, 0x80, 0x80, 0x80,
+    0x80, 0x80, 0x80, 0x80, 0x80, 0x80, 0x80, 0x80, 0x80, 0x80, 0x80, 0x80, 0x80, 0x80, 0x80, 0x80,
+    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00
+    ], [
+    0x00, 0x7e, 0x7c, 0x7c, 0x78, 0x78, 0x78, 0x78, 0x70, 0x70, 0x70, 0x70, 0x70, 0x70, 0x70, 0x70,
+    0x60, 0x60, 0x60, 0x60, 0x60, 0x60, 0x60, 0x60, 0x60, 0x60, 0x60, 0x60, 0x60, 0x60, 0x60, 0x60,
+    0x40, 0x40, 0x40, 0x40, 0x40, 0x40, 0x40, 0x40, 0x40, 0x40, 0x40, 0x40, 0x40, 0x40, 0x40, 0x40,
+    0x40, 0x40, 0x40, 0x40, 0x40, 0x40, 0x40, 0x40, 0x40, 0x40, 0x40, 0x40, 0x40, 0x40, 0x40, 0x40,
+    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+    0x40, 0x7e, 0x7c, 0x7c, 0x78, 0x78, 0x78, 0x78, 0x70, 0x70, 0x70, 0x70, 0x70, 0x70, 0x70, 0x70,
+    0x60, 0x60, 0x60, 0x60, 0x60, 0x60, 0x60, 0x60, 0x60, 0x60, 0x60, 0x60, 0x60, 0x60, 0x60, 0x60,
+    0x40, 0x40, 0x40, 0x40, 0x40, 0x40, 0x40, 0x40, 0x40, 0x40, 0x40, 0x40, 0x40, 0x40, 0x40, 0x40,
+    0x40, 0x40, 0x40, 0x40, 0x40, 0x40, 0x40, 0x40, 0x40, 0x40, 0x40, 0x40, 0x40, 0x40, 0x40, 0x40,
+    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00
+    ], [
+    0x00, 0x3e, 0x3c, 0x3c, 0x38, 0x38, 0x38, 0x38, 0x30, 0x30, 0x30, 0x30, 0x30, 0x30, 0x30, 0x30,
+    0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20,
+    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+    0x20, 0x3e, 0x3c, 0x3c, 0x38, 0x38, 0x38, 0x38, 0x30, 0x30, 0x30, 0x30, 0x30, 0x30, 0x30, 0x30,
+    0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20,
+    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+    0x60, 0x7e, 0x7c, 0x7c, 0x78, 0x78, 0x78, 0x78, 0x70, 0x70, 0x70, 0x70, 0x70, 0x70, 0x70, 0x70,
+    0x60, 0x60, 0x60, 0x60, 0x60, 0x60, 0x60, 0x60, 0x60, 0x60, 0x60, 0x60, 0x60, 0x60, 0x60, 0x60,
+    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+    0x20, 0x3e, 0x3c, 0x3c, 0x38, 0x38, 0x38, 0x38, 0x30, 0x30, 0x30, 0x30, 0x30, 0x30, 0x30, 0x30,
+    0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20,
+    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00
+    ], [
+    0x00, 0x1e, 0x1c, 0x1c, 0x18, 0x18, 0x18, 0x18, 0x10, 0x10, 0x10, 0x10, 0x10, 0x10, 0x10, 0x10,
+    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+    0x10, 0x1e, 0x1c, 0x1c, 0x18, 0x18, 0x18, 0x18, 0x10, 0x10, 0x10, 0x10, 0x10, 0x10, 0x10, 0x10,
+    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+    0x30, 0x3e, 0x3c, 0x3c, 0x38, 0x38, 0x38, 0x38, 0x30, 0x30, 0x30, 0x30, 0x30, 0x30, 0x30, 0x30,
+    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+    0x10, 0x1e, 0x1c, 0x1c, 0x18, 0x18, 0x18, 0x18, 0x10, 0x10, 0x10, 0x10, 0x10, 0x10, 0x10, 0x10,
+    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+    0x70, 0x7e, 0x7c, 0x7c, 0x78, 0x78, 0x78, 0x78, 0x70, 0x70, 0x70, 0x70, 0x70, 0x70, 0x70, 0x70,
+    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+    0x10, 0x1e, 0x1c, 0x1c, 0x18, 0x18, 0x18, 0x18, 0x10, 0x10, 0x10, 0x10, 0x10, 0x10, 0x10, 0x10,
+    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+    0x30, 0x3e, 0x3c, 0x3c, 0x38, 0x38, 0x38, 0x38, 0x30, 0x30, 0x30, 0x30, 0x30, 0x30, 0x30, 0x30,
+    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+    0x10, 0x1e, 0x1c, 0x1c, 0x18, 0x18, 0x18, 0x18, 0x10, 0x10, 0x10, 0x10, 0x10, 0x10, 0x10, 0x10,
+    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00
+    ], [
+    0x00, 0x0e, 0x0c, 0x0c, 0x08, 0x08, 0x08, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+    0x08, 0x0e, 0x0c, 0x0c, 0x08, 0x08, 0x08, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+    0x18, 0x1e, 0x1c, 0x1c, 0x18, 0x18, 0x18, 0x18, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+    0x08, 0x0e, 0x0c, 0x0c, 0x08, 0x08, 0x08, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+    0x38, 0x3e, 0x3c, 0x3c, 0x38, 0x38, 0x38, 0x38, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+    0x08, 0x0e, 0x0c, 0x0c, 0x08, 0x08, 0x08, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+    0x18, 0x1e, 0x1c, 0x1c, 0x18, 0x18, 0x18, 0x18, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+    0x08, 0x0e, 0x0c, 0x0c, 0x08, 0x08, 0x08, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+    0x78, 0x7e, 0x7c, 0x7c, 0x78, 0x78, 0x78, 0x78, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+    0x08, 0x0e, 0x0c, 0x0c, 0x08, 0x08, 0x08, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+    0x18, 0x1e, 0x1c, 0x1c, 0x18, 0x18, 0x18, 0x18, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+    0x08, 0x0e, 0x0c, 0x0c, 0x08, 0x08, 0x08, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+    0x38, 0x3e, 0x3c, 0x3c, 0x38, 0x38, 0x38, 0x38, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+    0x08, 0x0e, 0x0c, 0x0c, 0x08, 0x08, 0x08, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+    0x18, 0x1e, 0x1c, 0x1c, 0x18, 0x18, 0x18, 0x18, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+    0x08, 0x0e, 0x0c, 0x0c, 0x08, 0x08, 0x08, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00
+    ], [
+    0x00, 0x06, 0x04, 0x04, 0x00, 0x00, 0x00, 0x00, 0x04, 0x06, 0x04, 0x04, 0x00, 0x00, 0x00, 0x00,
+    0x0c, 0x0e, 0x0c, 0x0c, 0x00, 0x00, 0x00, 0x00, 0x04, 0x06, 0x04, 0x04, 0x00, 0x00, 0x00, 0x00,
+    0x1c, 0x1e, 0x1c, 0x1c, 0x00, 0x00, 0x00, 0x00, 0x04, 0x06, 0x04, 0x04, 0x00, 0x00, 0x00, 0x00,
+    0x0c, 0x0e, 0x0c, 0x0c, 0x00, 0x00, 0x00, 0x00, 0x04, 0x06, 0x04, 0x04, 0x00, 0x00, 0x00, 0x00,
+    0x3c, 0x3e, 0x3c, 0x3c, 0x00, 0x00, 0x00, 0x00, 0x04, 0x06, 0x04, 0x04, 0x00, 0x00, 0x00, 0x00,
+    0x0c, 0x0e, 0x0c, 0x0c, 0x00, 0x00, 0x00, 0x00, 0x04, 0x06, 0x04, 0x04, 0x00, 0x00, 0x00, 0x00,
+    0x1c, 0x1e, 0x1c, 0x1c, 0x00, 0x00, 0x00, 0x00, 0x04, 0x06, 0x04, 0x04, 0x00, 0x00, 0x00, 0x00,
+    0x0c, 0x0e, 0x0c, 0x0c, 0x00, 0x00, 0x00, 0x00, 0x04, 0x06, 0x04, 0x04, 0x00, 0x00, 0x00, 0x00,
+    0x7c, 0x7e, 0x7c, 0x7c, 0x00, 0x00, 0x00, 0x00, 0x04, 0x06, 0x04, 0x04, 0x00, 0x00, 0x00, 0x00,
+    0x0c, 0x0e, 0x0c, 0x0c, 0x00, 0x00, 0x00, 0x00, 0x04, 0x06, 0x04, 0x04, 0x00, 0x00, 0x00, 0x00,
+    0x1c, 0x1e, 0x1c, 0x1c, 0x00, 0x00, 0x00, 0x00, 0x04, 0x06, 0x04, 0x04, 0x00, 0x00, 0x00, 0x00,
+    0x0c, 0x0e, 0x0c, 0x0c, 0x00, 0x00, 0x00, 0x00, 0x04, 0x06, 0x04, 0x04, 0x00, 0x00, 0x00, 0x00,
+    0x3c, 0x3e, 0x3c, 0x3c, 0x00, 0x00, 0x00, 0x00, 0x04, 0x06, 0x04, 0x04, 0x00, 0x00, 0x00, 0x00,
+    0x0c, 0x0e, 0x0c, 0x0c, 0x00, 0x00, 0x00, 0x00, 0x04, 0x06, 0x04, 0x04, 0x00, 0x00, 0x00, 0x00,
+    0x1c, 0x1e, 0x1c, 0x1c, 0x00, 0x00, 0x00, 0x00, 0x04, 0x06, 0x04, 0x04, 0x00, 0x00, 0x00, 0x00,
+    0x0c, 0x0e, 0x0c, 0x0c, 0x00, 0x00, 0x00, 0x00, 0x04, 0x06, 0x04, 0x04, 0x00, 0x00, 0x00, 0x00
+    ], [
+    0x00, 0x02, 0x00, 0x00, 0x02, 0x02, 0x00, 0x00, 0x06, 0x06, 0x00, 0x00, 0x02, 0x02, 0x00, 0x00,
+    0x0e, 0x0e, 0x00, 0x00, 0x02, 0x02, 0x00, 0x00, 0x06, 0x06, 0x00, 0x00, 0x02, 0x02, 0x00, 0x00,
+    0x1e, 0x1e, 0x00, 0x00, 0x02, 0x02, 0x00, 0x00, 0x06, 0x06, 0x00, 0x00, 0x02, 0x02, 0x00, 0x00,
+    0x0e, 0x0e, 0x00, 0x00, 0x02, 0x02, 0x00, 0x00, 0x06, 0x06, 0x00, 0x00, 0x02, 0x02, 0x00, 0x00,
+    0x3e, 0x3e, 0x00, 0x00, 0x02, 0x02, 0x00, 0x00, 0x06, 0x06, 0x00, 0x00, 0x02, 0x02, 0x00, 0x00,
+    0x0e, 0x0e, 0x00, 0x00, 0x02, 0x02, 0x00, 0x00, 0x06, 0x06, 0x00, 0x00, 0x02, 0x02, 0x00, 0x00,
+    0x1e, 0x1e, 0x00, 0x00, 0x02, 0x02, 0x00, 0x00, 0x06, 0x06, 0x00, 0x00, 0x02, 0x02, 0x00, 0x00,
+    0x0e, 0x0e, 0x00, 0x00, 0x02, 0x02, 0x00, 0x00, 0x06, 0x06, 0x00, 0x00, 0x02, 0x02, 0x00, 0x00,
+    0x7e, 0x7e, 0x00, 0x00, 0x02, 0x02, 0x00, 0x00, 0x06, 0x06, 0x00, 0x00, 0x02, 0x02, 0x00, 0x00,
+    0x0e, 0x0e, 0x00, 0x00, 0x02, 0x02, 0x00, 0x00, 0x06, 0x06, 0x00, 0x00, 0x02, 0x02, 0x00, 0x00,
+    0x1e, 0x1e, 0x00, 0x00, 0x02, 0x02, 0x00, 0x00, 0x06, 0x06, 0x00, 0x00, 0x02, 0x02, 0x00, 0x00,
+    0x0e, 0x0e, 0x00, 0x00, 0x02, 0x02, 0x00, 0x00, 0x06, 0x06, 0x00, 0x00, 0x02, 0x02, 0x00, 0x00,
+    0x3e, 0x3e, 0x00, 0x00, 0x02, 0x02, 0x00, 0x00, 0x06, 0x06, 0x00, 0x00, 0x02, 0x02, 0x00, 0x00,
+    0x0e, 0x0e, 0x00, 0x00, 0x02, 0x02, 0x00, 0x00, 0x06, 0x06, 0x00, 0x00, 0x02, 0x02, 0x00, 0x00,
+    0x1e, 0x1e, 0x00, 0x00, 0x02, 0x02, 0x00, 0x00, 0x06, 0x06, 0x00, 0x00, 0x02, 0x02, 0x00, 0x00,
+    0x0e, 0x0e, 0x00, 0x00, 0x02, 0x02, 0x00, 0x00, 0x06, 0x06, 0x00, 0x00, 0x02, 0x02, 0x00, 0x00
+    ], [
+    0x00, 0x00, 0x01, 0x00, 0x03, 0x00, 0x01, 0x00, 0x07, 0x00, 0x01, 0x00, 0x03, 0x00, 0x01, 0x00,
+    0x0f, 0x00, 0x01, 0x00, 0x03, 0x00, 0x01, 0x00, 0x07, 0x00, 0x01, 0x00, 0x03, 0x00, 0x01, 0x00,
+    0x1f, 0x00, 0x01, 0x00, 0x03, 0x00, 0x01, 0x00, 0x07, 0x00, 0x01, 0x00, 0x03, 0x00, 0x01, 0x00,
+    0x0f, 0x00, 0x01, 0x00, 0x03, 0x00, 0x01, 0x00, 0x07, 0x00, 0x01, 0x00, 0x03, 0x00, 0x01, 0x00,
+    0x3f, 0x00, 0x01, 0x00, 0x03, 0x00, 0x01, 0x00, 0x07, 0x00, 0x01, 0x00, 0x03, 0x00, 0x01, 0x00,
+    0x0f, 0x00, 0x01, 0x00, 0x03, 0x00, 0x01, 0x00, 0x07, 0x00, 0x01, 0x00, 0x03, 0x00, 0x01, 0x00,
+    0x1f, 0x00, 0x01, 0x00, 0x03, 0x00, 0x01, 0x00, 0x07, 0x00, 0x01, 0x00, 0x03, 0x00, 0x01, 0x00,
+    0x0f, 0x00, 0x01, 0x00, 0x03, 0x00, 0x01, 0x00, 0x07, 0x00, 0x01, 0x00, 0x03, 0x00, 0x01, 0x00,
+    0x7f, 0x00, 0x01, 0x00, 0x03, 0x00, 0x01, 0x00, 0x07, 0x00, 0x01, 0x00, 0x03, 0x00, 0x01, 0x00,
+    0x0f, 0x00, 0x01, 0x00, 0x03, 0x00, 0x01, 0x00, 0x07, 0x00, 0x01, 0x00, 0x03, 0x00, 0x01, 0x00,
+    0x1f, 0x00, 0x01, 0x00, 0x03, 0x00, 0x01, 0x00, 0x07, 0x00, 0x01, 0x00, 0x03, 0x00, 0x01, 0x00,
+    0x0f, 0x00, 0x01, 0x00, 0x03, 0x00, 0x01, 0x00, 0x07, 0x00, 0x01, 0x00, 0x03, 0x00, 0x01, 0x00,
+    0x3f, 0x00, 0x01, 0x00, 0x03, 0x00, 0x01, 0x00, 0x07, 0x00, 0x01, 0x00, 0x03, 0x00, 0x01, 0x00,
+    0x0f, 0x00, 0x01, 0x00, 0x03, 0x00, 0x01, 0x00, 0x07, 0x00, 0x01, 0x00, 0x03, 0x00, 0x01, 0x00,
+    0x1f, 0x00, 0x01, 0x00, 0x03, 0x00, 0x01, 0x00, 0x07, 0x00, 0x01, 0x00, 0x03, 0x00, 0x01, 0x00,
+    0x0f, 0x00, 0x01, 0x00, 0x03, 0x00, 0x01, 0x00, 0x07, 0x00, 0x01, 0x00, 0x03, 0x00, 0x01, 0x00
+    ]
+];
+
+/// Contains indexing information that takes a coordinate and produces x and y values for the move
+/// making function.
+pub const INDEXING: [(usize, usize, usize, usize); 64] = [
+    (0x00, 0x00, 0x00, 0x07), (0x01, 0x00, 0x01, 0x06),
+    (0x02, 0x00, 0x02, 0x05), (0x03, 0x00, 0x03, 0x04),
+    (0x04, 0x00, 0x04, 0x03), (0x05, 0x00, 0x05, 0x02),
+    (0x06, 0x00, 0x06, 0x01), (0x07, 0x00, 0x07, 0x00),
+    (0x00, 0x01, 0x01, 0x07), (0x01, 0x01, 0x01, 0x06),
+    (0x02, 0x01, 0x02, 0x05), (0x03, 0x01, 0x03, 0x04),
+    (0x04, 0x01, 0x04, 0x03), (0x05, 0x01, 0x05, 0x02),
+    (0x06, 0x01, 0x06, 0x01), (0x07, 0x01, 0x07, 0x01),
+    (0x00, 0x02, 0x02, 0x07), (0x01, 0x02, 0x02, 0x06),
+    (0x02, 0x02, 0x02, 0x05), (0x03, 0x02, 0x03, 0x04),
+    (0x04, 0x02, 0x04, 0x03), (0x05, 0x02, 0x05, 0x02),
+    (0x06, 0x02, 0x06, 0x02), (0x07, 0x02, 0x07, 0x02),
+    (0x00, 0x03, 0x03, 0x07), (0x01, 0x03, 0x03, 0x06),
+    (0x02, 0x03, 0x03, 0x05), (0x03, 0x03, 0x03, 0x04),
+    (0x04, 0x03, 0x04, 0x03), (0x05, 0x03, 0x05, 0x03),
+    (0x06, 0x03, 0x06, 0x03), (0x07, 0x03, 0x07, 0x03),
+    (0x00, 0x04, 0x04, 0x07), (0x01, 0x04, 0x04, 0x06),
+    (0x02, 0x04, 0x04, 0x05), (0x03, 0x04, 0x04, 0x04),
+    (0x04, 0x04, 0x04, 0x04), (0x05, 0x04, 0x05, 0x04),
+    (0x06, 0x04, 0x06, 0x04), (0x07, 0x04, 0x07, 0x04),
+    (0x00, 0x05, 0x05, 0x07), (0x01, 0x05, 0x05, 0x06),
+    (0x02, 0x05, 0x05, 0x05), (0x03, 0x05, 0x05, 0x05),
+    (0x04, 0x05, 0x05, 0x05), (0x05, 0x05, 0x05, 0x05),
+    (0x06, 0x05, 0x06, 0x05), (0x07, 0x05, 0x07, 0x05),
+    (0x00, 0x06, 0x06, 0x07), (0x01, 0x06, 0x06, 0x06),
+    (0x02, 0x06, 0x06, 0x06), (0x03, 0x06, 0x06, 0x06),
+    (0x04, 0x06, 0x06, 0x06), (0x05, 0x06, 0x06, 0x06),
+    (0x06, 0x06, 0x06, 0x06), (0x07, 0x06, 0x07, 0x06),
+    (0x00, 0x07, 0x07, 0x07), (0x01, 0x07, 0x07, 0x07),
+    (0x02, 0x07, 0x07, 0x07), (0x03, 0x07, 0x07, 0x07),
+    (0x04, 0x07, 0x07, 0x07), (0x05, 0x07, 0x07, 0x07),
+    (0x06, 0x07, 0x07, 0x07), (0x07, 0x07, 0x07, 0x07)
+];
+
 /// A function which allows shifting a negative number of spots to the right.
 /// # Arguments
 /// * `x`: The number to shift.
@@ -231,6 +525,124 @@ pub fn all_moves(player: u64, opponent: u64) -> u64 {
     }
 
     all_moves & !(player | opponent)
+}
+
+#[cfg(all(feature="bmi2", target_arch="x86_64"))]
+pub fn get_flip(pos: usize, player: u64, opponent: u64) -> u64 {
+    let masks = EXTRACT_RAYS[pos];
+    const NOT_EDGES_H : u64 = !0x8181818181818181;
+    const NOT_EDGES_V : u64 = !0xFF000000000000FF;
+    const NOT_EDGES_D : u64 = !0xFF818181818181FF;
+
+    let ph: u64 = pext(player, masks[0]);
+    let pv: u64 = pext(player, masks[1]);
+    let pd: u64 = pext(player, masks[2]);
+    let pa: u64 = pext(player, masks[3]);
+
+    let oh: u64 = pext(opponent, masks[0] & NOT_EDGES_H);
+    let ov: u64 = pext(opponent, masks[1] & NOT_EDGES_V);
+    let od: u64 = pext(opponent, masks[2] & NOT_EDGES_D);
+    let oa: u64 = pext(opponent, masks[3] & NOT_EDGES_D);
+
+    let (x, y, x2, y2) = INDEXING[pos];
+
+    let fh = FLIP[x][(OUTFLANK[x][oh as usize] & ph as u8) as usize];
+    let fv = FLIP[y][(OUTFLANK[y][ov as usize] & pv as u8) as usize];
+    let fd = FLIP[x2][(OUTFLANK[x2][od as usize] & pd as u8) as usize];
+    let fa = FLIP[y2][(OUTFLANK[y2][oa as usize] & pa as u8) as usize];
+
+    let mh = pdep(fh as u64, masks[0]);
+    let mv = pdep(fv as u64, masks[1]);
+    let md = pdep(fd as u64, masks[2]);
+    let ma = pdep(fa as u64, masks[3]);
+
+    (mh | mv | md | ma) as u64
+}
+
+#[cfg(all(feature="bmi2", target_arch="x86"))]
+pub fn get_flip(pos: usize, player: u64, opponent: u64) -> u64 {
+    let masks = EXTRACT_RAYS[pos];
+    const NOT_EDGES_H: u64 = !0x8181818181818181;
+    const NOT_EDGES_V: u64 = !0xFF000000000000FF;
+    const NOT_EDGES_D: u64 = !0xFF818181818181FF;
+
+    let ph0 = pext(player as u32, masks[0] as u32) as u64;
+    let pv0 = pext(player as u32, masks[1] as u32) as u64;
+    let pd0 = pext(player as u32, masks[2] as u32) as u64;
+    let pa0 = pext(player as u32, masks[3] as u32) as u64;
+
+    let ph1 = pext((player >> 32) as u32, (masks[0] >> 32) as u32) as u64;
+    let pv1 = pext((player >> 32) as u32, (masks[1] >> 32) as u32) as u64;
+    let pd1 = pext((player >> 32) as u32, (masks[2] >> 32) as u32) as u64;
+    let pa1 = pext((player >> 32) as u32, (masks[3] >> 32) as u32) as u64;
+
+    let oh0 = pext(opponent as u32, (masks[0] & NOT_EDGES_H) as u32) as u64;
+    let ov0 = pext(opponent as u32, (masks[1] & NOT_EDGES_V) as u32) as u64;
+    let od0 = pext(opponent as u32, (masks[2] & NOT_EDGES_D) as u32) as u64;
+    let oa0 = pext(opponent as u32, (masks[3] & NOT_EDGES_D) as u32) as u64;
+
+    let oh1 = pext((opponent >> 32) as u32, ((masks[0] & NOT_EDGES_H) >> 32) as u32) as u64;
+    let ov1 = pext((opponent >> 32) as u32, ((masks[1] & NOT_EDGES_V) >> 32) as u32) as u64;
+    let od1 = pext((opponent >> 32) as u32, ((masks[2] & NOT_EDGES_D) >> 32) as u32) as u64;
+    let oa1 = pext((opponent >> 32) as u32, ((masks[3] & NOT_EDGES_D) >> 32) as u32) as u64;
+
+    let (x, y, x2, y2) = INDEXING[pos];
+
+    const SHIFTS_P: [u64; 15] = [0,0,0,0,1,2,3,4,4,4,4,4,3,2,1];
+    const SHIFTS_O: [u64; 15] = [0,0,0,0,0,1,2,3,3,3,3,2,1,0,0];
+
+    let ph: u64 = ph0 | ph1;
+    let pv: u64 = pv0 | (pv1 << 4);
+    let pd: u64 = pd0 | (pd1 << SHIFTS_P[7 - x + y]);
+    let pa: u64 = pa0 | (pa1 << SHIFTS_P[x + y]);
+
+    let oh: u64 = oh0 | oh1;
+    let ov: u64 = ov0 | (ov1 << 3);
+    let od: u64 = od0 | (od1 << SHIFTS_O[7 - x + y]);
+    let oa: u64 = oa0 | (oa1 << SHIFTS_O[x + y]);
+
+    let fh = FLIP[x][(OUTFLANK[x][oh as usize] & ph as u8) as usize];
+    let fv = FLIP[y][(OUTFLANK[y][ov as usize] & pv as u8) as usize];
+    let fd = FLIP[x2][(OUTFLANK[x2][od as usize] & pd as u8) as usize];
+    let fa = FLIP[y2][(OUTFLANK[y2][oa as usize] & pa as u8) as usize];
+
+    let mh0 = pdep(fh as u32, masks[0] as u32) as u64;
+    let mv0 = pdep(fv as u32, masks[1] as u32) as u64;
+    let md0 = pdep(fd as u32, masks[2] as u32) as u64;
+    let ma0 = pdep(fa as u32, masks[3] as u32) as u64;
+
+    let mh1 = pdep(fh as u32,                                   (masks[0] >> 32) as u32) as u64;
+    let mv1 = pdep((fv as u64 >> 4) as u32,                     (masks[1] >> 32) as u32) as u64;
+    let md1 = pdep((fd as u64 >> SHIFTS_P[7 - x + y]) as u32,   (masks[2] >> 32) as u32) as u64;
+    let ma1 = pdep((fa as u64 >> SHIFTS_P[x + y]) as u32,       (masks[3] >> 32) as u32) as u64;
+
+    (mh0 | mv0 | md0 | ma0  |  (mh1 << 32) | (mv1 << 32) | (md1 << 32) | (ma1 << 32)) as u64
+}
+
+#[cfg(not(feature="bmi2"))]
+pub fn get_flip(pos: usize, player: u64, opponent: u64) -> u64 {
+    let disk = 0x80_00_00_00_00_00_00_00 >> pos;
+    let mut flood = 0;
+    let player = player | disk;
+    for i in 0..SHIFT_DIRS.len() {
+        let shift = SHIFT_DIRS[i];
+        let prop = opponent & SHIFT_MASKS[i] & SHIFT_RAYS[pos][i];
+        let mut temp_flood = 0;
+
+        let mut gen = disk;
+        let mut next = gen;
+        while gen != 0 {
+            temp_flood |= gen;
+            next = directional_shift(gen, shift);
+            gen = next & prop;
+        }
+
+        if next & player != 0 {
+            flood |= temp_flood ^ disk;
+        }
+    }
+
+    flood | disk
 }
 
 #[cfg(test)]
