+++ src/board/mod.rs
@@ -1,7 +1,6 @@
 use std::fmt;
 
 pub mod bitboard;
-pub mod do_moves_fast;
 
 #[cfg(test)]
 pub mod test;
@@ -159,34 +158,6 @@ impl Board {
         match move_option {
             Move::Play(m) => {
                 let disk = 0x80_00_00_00_00_00_00_00 >> m;
-                // let num_directions = bitboard::SHIFT_DIRS.len();
-
-                // let (player, opponent) = if self.black_move {
-                //     self.black_disks |= disk;
-                //     (self.black_disks, self.white_disks)
-                // } else {
-                //     self.white_disks |= disk;
-                //     (self.white_disks, self.black_disks)
-                // };
-                //
-                // let mut flood = 0;
-                // for i in 0..num_directions {
-                //     let shift = bitboard::SHIFT_DIRS[i];
-                //     let prop = opponent & bitboard::SHIFT_MASKS[i] & bitboard::SHIFT_RAYS[m as usize][i];
-                //     let mut temp_flood = 0;
-                //
-                //     let mut gen = disk;
-                //     let mut next = gen;
-                //     while gen != 0 {
-                //         temp_flood |= gen;
-                //         next = bitboard::directional_shift(gen, shift);
-                //         gen = next & prop;
-                //     }
-                //
-                //     if next & player != 0 {
-                //         flood |= temp_flood ^ disk;
-                //     }
-                // }
 
                 let (player, opponent) = if self.black_move {
                     (self.black_disks, self.white_disks)
@@ -194,7 +165,7 @@ impl Board {
                     (self.white_disks, self.black_disks)
                 };
 
-                let flood = do_moves_fast::do_move_no_asm(m as usize, player, opponent);
+                let flood = bitboard::get_flip(m as usize, player, opponent);
 
                 self.white_disks ^= flood;
                 self.black_disks ^= flood;
