+++ src/lib.rs
@@ -1,3 +1,4 @@
+#[cfg(feature="bmi2")]
 extern crate bitintr;
 
 pub mod board;
